+++ src/Model/Aurora.php
@@ -558,7 +558,6 @@ class Aurora
         $columns[] = $columnName;
         $values[] = ':' . $columnName;
         $params[$columnName] = $this->getPropertyValueForDatabase($property);
-        ;
     }
 
     /**