+++ CHANGELOG.md
@@ -5,6 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres 
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- N/A
+
+### Changed
+- `Aurora::with()` no longer takes a string array of FQCNs, instead uses the shape `array<string, array<int, string>>` 
+to allow fetching of further associated objects
+
+### Deprecated
+- N/A
+
+### Removed
+- N/A
+
+### Fixed
+- N/A
+
+### Security
+- N/A
+
+---
+
 ## [2.7.3] - 2024-08-27
 ### Added
 - N/A